kvutils: Move pre-execution write set selection into the committer.
There's really only one way to do it.
