+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/validator/preexecution/PreExecutingValidatingCommitter.scala
@@ -3,6 +3,8 @@
 
 package com.daml.ledger.validator.preexecution
 
+import java.time.Instant
+
 import com.daml.ledger.participant.state.kvutils.DamlKvutils.DamlStateKey
 import com.daml.ledger.participant.state.kvutils.Raw
 import com.daml.ledger.participant.state.v1.{ParticipantId, SubmissionResult}
@@ -19,12 +21,14 @@ import scala.util.{Failure, Success}
   * fingerprints alongside values), parametric in the logic that produces a fingerprint given a
   * value.
   *
+  * @param now                           Returns the current time.
   * @param transformStateReader          Transforms the state reader into the format used by the underlying store.
   * @param validator                     The pre-execution validator.
   * @param postExecutionConflictDetector The post-execution conflict detector.
-  * @param postExecutionFinalizer        The post-execution finalizer.
+  * @param postExecutionWriter           The post-execution writer.
   */
 class PreExecutingValidatingCommitter[StateValue, ReadSet, WriteSet](
+    now: () => Instant,
     transformStateReader: LedgerStateReader => StateReader[DamlStateKey, StateValue],
     validator: PreExecutingSubmissionValidator[StateValue, ReadSet, WriteSet],
     postExecutionConflictDetector: PostExecutionConflictDetector[
@@ -33,7 +37,7 @@ class PreExecutingValidatingCommitter[StateValue, ReadSet, WriteSet](
       ReadSet,
       WriteSet,
     ],
-    postExecutionFinalizer: PostExecutionFinalizer[ReadSet, WriteSet],
+    postExecutionWriter: PostExecutionWriter[WriteSet],
 ) {
 
   private val logger = ContextualizedLogger.get(getClass)
@@ -68,13 +72,25 @@ class PreExecutingValidatingCommitter[StateValue, ReadSet, WriteSet](
               Success(SubmissionResult.Acknowledged) // But it will simply be dropped.
             case result => result
           }
-          submissionResult <- postExecutionFinalizer.finalizeSubmission(
-            preExecutionOutput,
-            ledgerStateOperations,
-          )
+          writeSet = selectWriteSet(preExecutionOutput)
+          submissionResult <- postExecutionWriter.write(writeSet, ledgerStateOperations)
         } yield submissionResult
       }
     }
+
+  private def selectWriteSet(
+      preExecutionOutput: PreExecutionOutput[ReadSet, WriteSet]
+  ): WriteSet = {
+    val recordTime = now()
+    val withinTimeBounds =
+      !recordTime.isBefore(preExecutionOutput.minRecordTime.getOrElse(Instant.MIN)) &&
+        !recordTime.isAfter(preExecutionOutput.maxRecordTime.getOrElse(Instant.MAX))
+    if (withinTimeBounds) {
+      preExecutionOutput.successWriteSet
+    } else {
+      preExecutionOutput.outOfTimeBoundsWriteSet
+    }
+  }
 
   private[this] def retry: PartialFunction[Throwable, Boolean] => RetryStrategy =
     RetryStrategy.constant(attempts = Some(3), 5.seconds)
